Database Connection Fixed
I got the right information in the db_conn file, and I also put the PDO
creation into that file as well so $conn will already be initialized on
every page. And I went ahead and removed the redundant PDO creation from
my files. The user in db_conn only has limited permissions, so if you
have to create or alter any tables, use the login info for the account I
sent you.